feat: complete the bugfix
move test from select-query.test.ts to postgres/query-generator.test.js
to allow other dialects to run
